requirements.txt
torch>=1.9.0
transformers>=4.21.0
librosa>=0.8.0
numpy>=1.21.0
soundfile>=0.10.0
datasets>=2.0.0
accelerate>=0.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
rapidfuzz>=3.0.0
phonetics>=1.0.5
editdistance>=0.6.0
jellyfish>=0.9.0
elevenlabs>=1.0.0
pydub>=0.25.0
python-dotenv>=1.0.0
textgrids>=1.0.0
webrtcvad>=2.0.10 